requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic~=1.14.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
beautifulsoup4==4.12.3
boto3==1.35.87
botocore==1.35.87
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.15
dirtyjson==1.0.8
distro==1.9.0
fastapi==0.115.6
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.13
langchain-community~=0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langsmith==0.2.6
llama-cloud==0.1.7
llama-index==0.12.8
llama-index-agent-openai==0.4.1
llama-index-cli==0.4.0
llama-index-core==0.12.8
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-llms-openai==0.3.12
llama-index-multi-modal-llms-openai==0.4.1
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.1
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.18
marshmallow==3.23.2
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy~=1.26.4
openai==1.58.1
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic~=2.9.2
pydantic_core~=2.23.4
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-magic~=0.4.27
python-multipart==0.0.18
pytz==2024.2
PyYAML==6.0.2
redis
sentence-transformers
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.3
striprtf==0.0.26
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn~=0.34.0
wrapt==1.17.0
yarl==1.18.3

unstructured~=0.16.11
html5lib~=1.1
lxml~=5.3.0
olefile~=0.47
pip~=24.3.1
emoji~=2.14.0
cryptography~=44.0.0
cffi~=1.17.1
Mako~=1.3.8
MarkupSafe~=3.0.2
psutil~=6.1.1
backoff~=2.2.1
chardet~=5.2.0
webencodings~=0.5.1
pycparser~=2.22
RapidFuzz~=3.11.0
langdetect~=1.0.9
eval_type_backport~=0.2.2
unstructured-client~=0.28.1
python-iso639~=2024.10.22
aiofiles~=24.1.0
unstructured[pdf]

Booktype~=1.5
setuptools~=75.6.0
simplejson~=3.19.3
pydantic-settings~=2.7.0